requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
qdrant-client==1.9.0
sentence-transformers==3.0.1
numpy==1.26.4
aiosqlite==0.20.0
pypdf==4.3.1
python-dotenv==1.0.1
requests
openai==1.51.0
httpx==0.26.0

